% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI.R
\name{BMI}
\alias{BMI}
\title{BMI}
\usage{
BMI(waga, wzrost, plec = c("kobieta", "mezczyzna"))
}
\arguments{
\item{waga}{numeric, waga osoby, dla której liczymy BMI wyrażona w kg}

\item{wzrost}{numeric, wzrost osoby, dla której liczymy BMI wyrażony w metrach}

\item{plec}{string, plec osoby, dla której liczymy BMI. Możliwe wartości dla parametru: \code{kobieta}, \code{mezczyzna}}
}
\description{
Funkcja pomocnicza do liczenia wskaźnika BMI.
}
\examples{
BMI(100, 1.7)
BMI(85, 1.9, "kobieta")

}
