% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_journals.R
\name{capture_journals}
\alias{capture_journals}
\title{Evaluate an R expression while collecting journals from completed futures}
\usage{
capture_journals(expr, substitute = TRUE, envir = parent.frame())
}
\arguments{
\item{expr}{The R expression to evaluate}

\item{substitute}{If TRUE, then \code{expr} is subtituted, otherwise not.}

\item{envir}{The environment where \code{expr} should be evaluated}
}
\value{
A list of \link[=journal]{FutureJournal}:s.
}
\description{
Evaluate an R expression while collecting journals from completed futures
}
\details{
This function evaluates an R expression and capture the journals
signaled by futures as they are completed. A future \link{journal} comprise
a log of events appearing during the life-span of a future, e.g.
the timestamps when the future was created, launched, queried,
resolved, and its results are collected.
}
\examples{
slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(multisession, workers = 2)
js <- capture_journals({
  fs <- lapply(3:1, FUN = function(x) future(slow_fcn(x)))
  value(fs)
})
print(js)

## Shut down parallel workers
plan(sequential)
}
