% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky_globals.R
\name{clusterExportSticky}
\alias{clusterExportSticky}
\title{Export globals to the sticky-globals environment of the cluster nodes}
\usage{
clusterExportSticky(cl, globals)
}
\arguments{
\item{cl}{(cluster) A cluster object as returned by
\code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}

\item{globals}{(list) A named list of sticky globals to be exported.}
}
\value{
(invisible; cluster) The cluster object.
}
\description{
Export globals to the sticky-globals environment of the cluster nodes
}
\details{
This requires that the \pkg{future} package is installed on the cluster
nodes.
}
\keyword{internals}
