% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFuture.R
\name{ClusterFuture-class}
\alias{ClusterFuture}
\alias{ClusterFuture-class}
\title{An cluster future is a future whose value will be resolved asynchroneously in a parallel process}
\usage{
ClusterFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  cluster, ...)
}
\arguments{
\item{object}{An R \link[base]{environment}.}

\item{\dots}{Not used.}
}
\value{
An object of class \code{ClusterFuture}.
}
\description{
An cluster future is a future whose value will be resolved asynchroneously in a parallel process
}
\seealso{
To evaluate an expression using "cluster future", see function
\code{\link{cluster}()}.
}
\keyword{internal}

