% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EagerFuture-class.R
\name{EagerFuture-class}
\alias{EagerFuture}
\alias{EagerFuture-class}
\title{A eager future represents a \link{Future} whose value will be resolved immediately}
\usage{
EagerFuture(object = new.env(parent = emptyenv()), ...)
}
\arguments{
\item{object}{An R \link[base]{environment}.}

\item{...}{Not used.}
}
\value{
An object of class \code{EagerFuture}.
}
\description{
A eager future represents a \link{Future} whose value will be resolved immediately
}
\seealso{
To evaluate an expression using "eager future", see function
\code{\link{eager}()}.
}

