% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_of_meal}
\alias{energy_of_meal}
\title{Total energy value of the meal}
\usage{
energy_of_meal(list_of_products, weight_of_products)
}
\arguments{
\item{list_of_products}{list, list of the names of the products added to the
preparation of the meal}

\item{weight_of_products}{list, list of weights of the products added to the
preparation of the meal}
}
\description{
Function for calculating the total energy value of a meal
}
\examples{
list_of_products <- list("Kawior", "Kawior")
weight_of_products <- list(50, 100)
energy_of_meal(list_of_products, weight_of_products)

}
