% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_of_product}
\alias{energy_of_product}
\title{Energy value of the product for the given weight}
\usage{
energy_of_product(product, weight)
}
\arguments{
\item{product}{character, name of used product}

\item{weight}{numeric, weight of the used product expressed in grams}
}
\description{
Function for calculating the energy value of a product for given weight
}
\examples{
energy_of_product("Kawior", 50)

}
