% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_total}
\alias{energy_total}
\title{Total energy value of the product}
\usage{
energy_total(weight_of_product, protein, fat, carbohydrates, fiber = 0)
}
\arguments{
\item{weight_of_product}{numeric, weight of used product}

\item{protein}{numeric, protein content in the product for which we calculate
the energy value expressed in grams}

\item{fat}{numeric, fat content in the product for which we calculate the
energy value expressed in grams}

\item{carbohydrates}{numeric, carbohydrates content in the product for which
we calculate the energy value expressed in grams}

\item{fiber}{numeric, fiber content in the product for which we calculate
the energy value expressed in grams, 0 by default}
}
\description{
Function for calculating the total energy value of a product
}
\examples{
energy_total(50, 1.2, 0.9, 11.8, 3)
energy_total(35, 8, 32, 52)


}
