% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_value_of_the_product_per_100}
\alias{energy_value_of_the_product_per_100}
\title{Energy value of the product per 100 grams}
\usage{
energy_value_of_the_product_per_100(protein, fat, carbohydrates, fiber = 0)
}
\arguments{
\item{protein}{numeric, protein content in the product for which we calculate
the energy value expressed in grams}

\item{fat}{numeric, fat content in the product for which we calculate the
energy value expressed in grams}

\item{carbohydrates}{numeric, carbohydrates content in the product for which
we calculate the energy value expressed in grams}

\item{fiber}{numeric, fiber content in the product for which we calculate
the energy value expressed in grams, 0 by default}
}
\description{
Function for calculating the energy value of a product per 100
grams
}
\examples{
energy_value_of_the_product_per_100(2.7, 1.5, 16)
energy_value_of_the_product_per_100(2.7, 1.5, 16, 2)

}
