% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterPSOCK.R
\name{find_rshcmd}
\alias{find_rshcmd}
\title{Search for SSH clients on the current system}
\usage{
find_rshcmd(which = NULL, must_work = TRUE)
}
\arguments{
\item{which}{A character vector specifying which types of SSH clients
to search for.  If NULL, a default set of clients supported by the
current platform is searched for.}

\item{must_work}{If TRUE and no clients was found, then an error
is produced, otherwise only a warning.}
}
\value{
A `rsh_caller` function, or `NULL`.
}
\description{
Search for SSH clients on the current system
}
\keyword{internal}
