% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flapply.R
\name{flapply}
\alias{flapply}
\title{Apply a Function over a List or Vector via Futures}
\usage{
flapply(x, FUN, ...)
}
\arguments{
\item{x}{A vector-like object to iterate over.}

\item{FUN}{A function taking at least one argument.}

\item{...}{(optional) Additional arguments pass to \code{FUN()}.}
}
\value{
A list.
}
\description{
Apply a Function over a List or Vector via Futures
}
\keyword{internal}

