% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Efficient fold / reduce / accumulate / combine of a vector}
\usage{
fold(x, f, left = TRUE, unname = TRUE, threshold = 1000L)
}
\arguments{
\item{x}{A vector.}

\item{f}{A binary function, i.e. a function take takes two arguments.}

\item{left}{If `TRUE`, vector is combined from the left (the first element),
otherwise the right (the last element).}

\item{unname}{If `TRUE`, function `f` is called as `f(unname(y), x[[ii]])`,
otherwise as `f(y, x[[ii]])`, which may introduce name `"y"`.}

\item{threshold}{An integer (>= 2) specifying the length where the
recursive divide'and'conquer call will stop and incremental building of
the partial value is performed.}
}
\value{
A vector.
}
\description{
Efficient fold / reduce / accumulate / combine of a vector
}
\details{
This function is a more efficient (memory and speed) of
[`base::Reduce(f, x, right = !left, accumulate = FALSE)`][base::Reduce],
especially when `x` is long.
}
