% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_api-ClusterFutureBackend-class.R,
%   R/backend_api-FutureBackend-class.R,
%   R/backend_api-MulticoreFutureBackend-class.R,
%   R/backend_api-MultisessionFutureBackend-class.R,
%   R/backend_api-SequentialFutureBackend-class.R
\name{ClusterFutureBackend}
\alias{ClusterFutureBackend}
\alias{FutureBackend}
\alias{launchFuture}
\alias{MulticoreFutureBackend}
\alias{MultisessionFutureBackend}
\alias{SequentialFutureBackend}
\title{A ClusterFutureBackend resolves futures in parallel using any PSOCK cluster}
\usage{
ClusterFutureBackend(
  workers = availableWorkers(),
  persistent = FALSE,
  earlySignal = TRUE,
  gc = FALSE,
  ...
)

FutureBackend(...)

launchFuture(backend, future, ...)

MulticoreFutureBackend(
  workers = availableCores(constraints = "multicore"),
  ...
)

MultisessionFutureBackend(workers = availableCores(), ...)

SequentialFutureBackend(...)
}
\arguments{
\item{workers}{...}

\item{persistent}{(deprecated) ...}

\item{backend}{a \link{FutureBackend}.}

\item{future}{a \link{Future} to be launched on the backend.}

\item{\ldots}{(optional) not used.}
}
\value{
\code{FutureBackend()} returns a \code{FutureBackend} that inherits an
environment.

\code{launchFuture()} returns a the launched \code{Future} object.
}
\description{
A ClusterFutureBackend resolves futures in parallel using any PSOCK cluster

A backend that controls how futures are launched on parallel workers

\code{launchFuture()} launches a future on the backend.

A MulticoreFutureBackend resolves futures in parallel using forked processing on the current machine

A MultisessionFutureBackend resolves futures in parallel using a PSOCK cluster on the current machine

A SequentialFutureBackend resolves futures sequentially in the current R session
}
\details{
The \code{ClusterFutureBackend} is selected by
\code{plan(cluster, workers = workers)}.

The \code{MulticoreFutureBackend} backend is selected by
\code{plan(multicore, workers = workers)}.

The \code{MultisessionFutureBackend} backend is selected by
\code{plan(multisession, workers = workers)}.

The \code{SequentialFutureBackend} is selected by \code{plan(sequential)}.
}
\keyword{internal}
