% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureBackendResources.R
\name{FutureBackendResources}
\alias{FutureBackendResources}
\title{Creates a backend resource object}
\usage{
FutureBackendResources()
}
\value{
Returns an environment of class \code{FutureBackendResources}.
}
\description{
Creates a backend resource object
}
\keyword{internal}
