% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureError-class.R
\name{FutureCondition}
\alias{FutureCondition}
\alias{getOutput}
\alias{FutureMessage}
\alias{FutureWarning}
\alias{FutureError}
\title{A condition (message, warning, or error) related to a future}
\usage{
FutureCondition(message, call = NULL, future = NULL, output = NULL)

FutureMessage(message, call = NULL, future = NULL, output = NULL)

FutureWarning(message, call = NULL, future = NULL, output = NULL)

FutureError(message, call = NULL, future = NULL, output = NULL)
}
\arguments{
\item{message}{A message.}

\item{call}{The call stack that led up to the condition.}

\item{future}{The \link{Future} involved.}

\item{output}{Output captured when condition occurred.}
}
\value{
An object of class FutureCondition which inherits from class
\link[base:conditions]{condition} and FutureMessage, FutureWarning,
and FutureError all inherits from FutureCondition.
Moreover, a FutureError inherits from \link[base:conditions]{error},
a FutureWarning from \link[base:conditions]{warning}, and
a FutureMessage from \link[base:conditions]{warning}.
}
\description{
A condition (message, warning, or error) related to a future
}
\keyword{internal}
