% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureError-class.R
\name{FutureError}
\alias{FutureError}
\alias{getOutput}
\title{An error occurred while trying to evaluate a future}
\usage{
FutureError(message, call = NULL, future = NULL, output = NULL)
}
\arguments{
\item{message}{The error message.}

\item{call}{The call stack that led up to the error.}

\item{future}{The \link{Future} involved.}

\item{output}{Output captured when error occurred.}
}
\value{
An \link[base:conditions]{error} object of class FutureError.
}
\description{
An error occurred while trying to evaluate a future
}
\keyword{internal}
