% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureEvaluationCondition-class.R
\name{FutureEvaluationCondition}
\alias{FutureEvaluationCondition}
\alias{getOutput}
\alias{FutureEvaluationMessage}
\alias{FutureEvaluationWarning}
\alias{FutureEvaluationError}
\title{A condition (message, warning, or error) that occurred while evaluation a future}
\usage{
FutureEvaluationCondition(message, call = NULL, future = NULL,
  output = NULL)

FutureEvaluationMessage(message, call = NULL, future = NULL,
  output = NULL)

FutureEvaluationWarning(message, call = NULL, future = NULL,
  output = NULL)

FutureEvaluationError(message, call = NULL, future = NULL, output = NULL)
}
\arguments{
\item{message}{A message.}

\item{call}{The call stack that led up to the condition.}

\item{future}{The \link{Future} involved.}

\item{output}{Output captured when condition occurred.}
}
\value{
An object of class FutureEvaluationCondition which inherits from class
\link[base:conditions]{condition} and FutureEvaluationMessage, FutureEvaluationWarning,
and FutureEvaluationError all inherits from FutureEvaluationCondition.
Moreover, a FutureEvaluationError inherits from \link[base:conditions]{error},
a FutureEvaluationWarning from \link[base:conditions]{warning}, and
a FutureEvaluationMessage from \link[base:conditions]{message}.
}
\description{
A condition (message, warning, or error) that occurred while evaluation a future
}
\keyword{internal}
