% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureGlobals-class.R
\name{FutureGlobals}
\alias{FutureGlobals}
\alias{as.FutureGlobals}
\alias{as.FutureGlobals.FutureGlobals}
\alias{as.FutureGlobals.Globals}
\alias{as.FutureGlobals.list}
\alias{[.FutureGlobals}
\alias{c.FutureGlobals}
\alias{unique.FutureGlobals}
\title{A representation of a set of globals used with futures}
\usage{
FutureGlobals(object = list(), resolved = FALSE, total_size = NA_real_, ...)
}
\arguments{
\item{object}{A named list.}

\item{resolved}{A logical indicating whether these globals
have been scanned for and resolved futures or not.}

\item{total_size}{The total size of all globals, if known.}

\item{\dots}{Not used.}
}
\value{
An object of class \code{FutureGlobals}.
}
\description{
A representation of a set of globals used with futures
}
\details{
This class extends the \link[globals]{Globals} class by add
attribute \code{resolved}.
}
\keyword{internal}
