% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/futureOf.R
\name{futureOf}
\alias{futureOf}
\title{Get the future of a future variable}
\usage{
futureOf(var = NULL, envir = parent.frame(), mustExist = TRUE,
  default = NA_character_)
}
\arguments{
\item{var}{the variable.  If NULL, all futures in the
environment are returned.}

\item{envir}{the environment where to search from.}

\item{mustExist}{If TRUE and the variable does not exists, then
an informative error is thrown, otherwise NA is returned.}

\item{default}{the default value if future was not found.}
}
\value{
A \link{Future} (or NA).  If \code{var} is NULL, then a
named list of Future:s are returned.
}
\description{
Get the future of a future variable
}

