% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureResult-class.R
\name{FutureResult}
\alias{FutureResult}
\alias{as.character.FutureResult}
\title{Results from resolving a future}
\usage{
FutureResult(value = NULL, stdout = NULL, conditions = NULL, ...,
  version = "1.7")

\method{as.character}{FutureResult}(x, ...)
}
\arguments{
\item{value}{The value of the future expression.
If the expression was not fully resolved (e.g. an error) occurred,
the the value is \code{NULL}.}

\item{conditions}{A list of zero or more list elements each containing
a captured [[base::condition]] and possibly more meta data such as the
call stack and a timestamp.}

\item{version}{The version format of the results.}

\item{\ldots}{(optional) Additional named results to be returned.}
}
\value{
An object of class FutureResult.
}
\description{
Results from resolving a future
}
\details{
This function is only part of the _backend_ Future API.
This function is _not_ part of the frontend Future API.
}
\section{Note to developers}{

The FutureResult structure is _under development_ and may change at anytime,
e.g. elements may be renamed or removed.  Because of this, please avoid
accessing the elements directly in code.  Feel free to reach out if you need
to do so in your code.
}

\section{Deprecated elements}{

* future (>= 1.11.0): elements \code{condition} and \code{calls} are
  deprecated in favor of \code{conditions}
}

\keyword{internal}
