% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protected_api-futures.R
\name{futures}
\alias{futures}
\title{Get all futures in a container}
\usage{
futures(x, ...)
}
\arguments{
\item{x}{An environment, a list, or a list environment.}

\item{\ldots}{Not used.}
}
\value{
An object of same type as \code{x} and with the same names
and/or dimensions, if set.
}
\description{
Gets all futures in an environment, a list, or a list environment
and returns an object of the same class (and dimensions).
Non-future elements are returned as is.
}
\details{
This function is useful for retrieve futures that were created via
future assignments (\verb{\%<-\%}) and therefore stored as promises.
This function turns such promises into standard \code{Future}
objects.
}
