% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureSessionInfo.R
\name{futureSessionInfo}
\alias{futureSessionInfo}
\title{Get session information specific to the future framework}
\usage{
futureSessionInfo(test = TRUE)
}
\arguments{
\item{test}{If TRUE, one or more futures are created to query workers
and validate their information.}
}
\value{
Nothing.
}
\description{
Get session information specific to the future framework
}
