% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize.R
\name{futurize}
\alias{futurize}
\title{Turns a function into a futuristic function}
\usage{
futurize(FUN, vectorize = FALSE)
}
\arguments{
\item{FUN}{A function.}
}
\value{
A function of class Futuristic
}
\description{
A futuristic function is a function that will return a future
function call such that when the value of the future is resolved
the value of the function call with be returned.
}
\seealso{
\code{\link{futureCall}()}
}

