% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycemic_index.R
\name{glycemic_index}
\alias{glycemic_index}
\title{Glycemic index of product}
\usage{
glycemic_index(product, weight)
}
\arguments{
\item{product}{string, name of used product}

\item{weight}{numeric, weight of used product}
}
\description{
Function for calculating glycemic index of product in prepared meal
}
\examples{
glycemic_index("Banan", 100)

}
