% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycemic_index.R
\name{glycemic_index_of_meal}
\alias{glycemic_index_of_meal}
\title{Glycemic index of prepared meal}
\usage{
glycemic_index_of_meal(list_of_products, weight_of_products)
}
\arguments{
\item{list_of_products}{list, list of the names of the products used to prepare the meal}

\item{weight_of_products}{list, list of weights of the products used to prepare the meal}
}
\description{
Function for calculating glycemic index of prepared meal
}
\examples{
list_of_products <- list("Banan", "Truskawki")
weight_of_products <- list(100, 100)
glycemic_index_of_meal(list_of_products, weight_of_products)

}
