% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_OP.R
\name{\%conditions\%}
\alias{\%conditions\%}
\title{Control whether standard output should be captured or not}
\usage{
fassignment \%conditions\% capture
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{capture}{If TRUE, the standard output will be captured, otherwise not.}
}
\description{
Control whether standard output should be captured or not
}
