% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_OP.R
\name{\%globals\%}
\alias{\%globals\%}
\title{Specify globals for a future assignment}
\usage{
fassignment \%globals\% globals
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<=\% \{ expr \}}.}

\item{globals}{A named list specifying globals to be used.}
}
\description{
Specify globals for a future assignment
}
\details{
`a %globals% b` is short for `a %tweak% list(globals = b)`.
}

