% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_OP.R
\name{\%label\%}
\alias{\%label\%}
\title{Specify label for a future assignment}
\usage{
fassignment \%label\% label
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{label}{An optional character string label attached to the future.}
}
\description{
Specify label for a future assignment
}
