% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_OP.R
\name{\%lazy\%}
\alias{\%lazy\%}
\title{Control lazy / eager evaluation for a future assignment}
\usage{
fassignment \%lazy\% value
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{value}{If \code{TRUE}, the future is resolved using lazy evaluation, otherwise eager evaluation.}
}
\description{
Control lazy / eager evaluation for a future assignment
}

