% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_OP.R
\name{\%plan\%}
\alias{\%plan\%}
\title{Use a specific plan for a future assignment}
\usage{
fassignment \%plan\% strategy
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{strategy}{The mechanism for how the future should be
resolved. See \code{\link{plan}()} for further details.}
}
\description{
Use a specific plan for a future assignment
}
\seealso{
The \code{\link{plan}()} function sets the default plan for all futures.
}
