% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api-interrupt.R
\name{interrupt}
\alias{interrupt}
\title{Interrupts a running future}
\usage{
interrupt(x, ...)
}
\arguments{
\item{x}{A Future.}

\item{\ldots}{All arguments used by the S3 methods.}
}
\value{
\code{interrupt()} returns the \link{Future} flagged as "interrupted",
if the backend supports interrupting futures.
}
\description{
Attempts to interrupt a running future. If the backend does
not support interrupting futures, nothing is done.
}
