% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LazyFuture.R
\name{LazyFuture-class}
\alias{LazyFuture}
\alias{LazyFuture-class}
\title{A lazy future represents a \link{Future} whose value will be resolved at the time when it is requested}
\usage{
LazyFuture(object = new.env(parent = emptyenv()), ...)
}
\arguments{
\item{object}{An R \link[base]{environment}.}

\item{...}{Not used.}
}
\value{
An object of class \code{LazyFuture}.
}
\description{
A lazy future represents a \link{Future} whose value will be resolved at the time when it is requested
}
\seealso{
To evaluate an expression using "lazy future", see function
\code{\link{lazy}()}.
%% Alternatively, use the infix assignment operator \code{\link{\%<-\%}}.
}

