% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{macro_percent_graph}
\alias{macro_percent_graph}
\title{Pie chart of percent of macronutrients in the prepared meal}
\usage{
macro_percent_graph(macronutrients_df)
}
\arguments{
\item{macronutrients_df}{data frame, data frame of amount of macronutrients used to prepare the meal}
}
\description{
Function for creating pie chart of percent of macronutrients in prepared meal
}
\examples{
list_of_products <- list("Kawior", "Agrest")
weight_of_products <- list(50, 100)
example_df <- macronutrients_of_meal(list_of_products, weight_of_products)
macro_percent_graph(example_df)

}
