% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_pie_chart.R
\name{macro_pie_chart}
\alias{macro_pie_chart}
\title{Interactive pie chart of percent of macronutrients in the prepared meal}
\usage{
macro_pie_chart(macro_percent)
}
\arguments{
\item{macro_percent}{data frame, data frame of amount of macronutrients used to prepare the meal}
}
\description{
Function for creating interactive pie chart of percent of macronutrients in prepared meal
}
\examples{
list_of_products <- list("Banan", "Truskawki")
weight_of_products <- list(100, 100)
macro_percent <- macronutrients_of_meal(list_of_products, weight_of_products)
macro_pie_chart(macro_percent)

}
