% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{macronutrients}
\alias{macronutrients}
\title{Amount of macronutrients in product}
\usage{
macronutrients(product, weight)
}
\arguments{
\item{product}{string, name of used product}

\item{weight}{numeric, weight of used product}
}
\description{
Function for calculating amount of macronutrients in product
}
\examples{
macronutrients("Kawior", 50)

}
