% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mainModuleUI}
\alias{mainModuleUI}
\title{Ui part of module for calculating energy of meal}
\usage{
mainModuleUI(id)
}
\arguments{
\item{id}{character, shiny id}
}
\description{
Ui part of module for calculating energy of meal
}
