% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCluster2.R
\name{makeCluster2}
\alias{makeCluster2}
\title{Create a Parallel Socket Cluster}
\usage{
makeCluster2(spec, user = NULL, rshopts = NULL, master = NULL,
  revtunnel = FALSE, ..., verbose = FALSE)
}
\arguments{
\item{spec}{Specification of the workers, e.g. number of background sessions or a vector of host names.}

\item{user}{The SSH user name.  If NULL, then it is not specified.}

\item{rshopts}{A character vector of optional SSH options.}

\item{master}{The host name or the IP number of the master, as known to the workers.  If NULL, then it defaults to \code{Sys.info()[["nodename"]]} unless \code{revtunnel = TRUE} in case it defaults to \code{"localhost"}.}

\item{revtunnel}{If TRUE, reverse SSH tunneling is used.}

\item{...}{Additional parameters passed to \code{\link[parallel:makeCluster]{makeCluster}()}.}

\item{verbose}{If TRUE, informative messages are outputted.}
}
\value{
An object of class \code{c("SOCKcluster", "cluster")}.
}
\description{
Creates a set of copies of \R running in parallel and communicating over sockets.
}
\examples{
\donttest{\dontrun{
cl <- makeCluster2("remote.server.org", rshopts = "-v", user = NULL,
                   revtunnel = TRUE, master = "localhost", port = 11001)
}}
}

