% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterMPI.R
\name{makeClusterMPI}
\alias{makeClusterMPI}
\title{Create an MPI cluster of \R workers for parallel processing}
\usage{
makeClusterMPI(workers, ..., verbose = getOption("future.debug", FALSE))
}
\arguments{
\item{workers}{The number workers (as a positive integer).}

\item{\dots}{Optional arguments passed to
\code{\link[parallel:makeCluster]{makeCluster(workers, type = "MPI", ...)}}.}

\item{verbose}{If TRUE, informative messages are outputted.}
}
\value{
An object of class \code{"FutureMPIcluster"} consisting
of a list of \code{"MPInode"} workers.
}
\description{
The \code{makeClusterMPI()} function creates a cluster of \R workers
for parallel processing.  This function works similarly to
\code{\link[parallel:makeCluster]{makeCluster(..., type = "MPI")}} of the
\pkg{parallel} package, but is a bit more user friendly and robust.
}
