% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_api-MulticoreFutureBackend-class.R
\name{MulticoreFuture-class}
\alias{MulticoreFuture-class}
\alias{resolved.MulticoreFuture}
\title{A multicore future is a future whose value will be resolved asynchronously in a parallel process}
\usage{
\method{resolved}{MulticoreFuture}(x, run = TRUE, timeout = NULL, ...)
}
\arguments{
\item{...}{Additional named elements passed to \code{\link[=Future]{Future()}}.}
}
\value{
\code{MulticoreFuture()} returns an object of class \code{MulticoreFuture}.
}
\description{
A multicore future is a future whose value will be resolved asynchronously in a parallel process
}
\section{Usage}{

To use 'multicore' futures, use \code{plan(multicore, ...)}, cf. \link{multicore}.
}

\keyword{internal}
