% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiprocessFuture-class.R
\name{MultiprocessFuture-class}
\alias{MultiprocessFuture}
\alias{MultiprocessFuture-class}
\title{An multiprocess future is a future whose value will be resolved asynchroneously in a parallel process}
\usage{
MultiprocessFuture(expr = NULL, envir = parent.frame(),
  substitute = FALSE, ...)
}
\arguments{
\item{expr}{An R \link[base]{expression}.}

\item{envir}{The \link{environment} in which the evaluation
is done (or inherits from if \code{local} is TRUE).}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{...}{Additional named elements of the future.}

\item{local}{If TRUE, the expression is evaluated such that
all assignments are done to local temporary environment, otherwise
the assignments are done to the global environment of the \R process
evaluating the future.}

\item{\dots}{Additional named elements passed to \code{\link{Future}()}.}
}
\value{
An object of class \code{MultiprocessFuture}.
}
\description{
An multiprocess future is a future whose value will be resolved asynchroneously in a parallel process
}
\seealso{
To evaluate an expression using "multiprocess future", see function
\code{\link{multiprocess}()}.
}
\keyword{internal}

