% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\title{The plan used to resolve the future}
\usage{
plan(strategy = NULL, ...)
}
\arguments{
\item{strategy}{The evaluation function to use for resolving a future.
If NULL, then the current strategy is returned.}

\item{...}{Additional arguments overriding the default arguments
of the evaluation function.}
}
\value{
If a new strategy is choosen, then the previous one is returned
(invisible), otherwise the current one is returned (visibly).
}
\description{
This function allows you to plan the future, more specifically,
it specifies how \code{\link{future}()}:s are resolved,
e.g. by eager or by lazy evaluation.
}
\examples{
plan(lazy)
f <- future({
  a <- 7
  b <- 3
  c <- 2
  a * b * c
})
y <- value(f)
print(y)


plan(eager)
f <- future({
  a <- 7
  b <- 3
  c <- 2
  a * b * c
})
y <- value(f)
print(y)
}
\seealso{
Evaluation functions provided by this package are
\code{\link{eager}()} and \code{\link{lazy}()}.
}

