% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.length}
\alias{.length}
\title{Gets the length of an object without dispatching}
\usage{
.length(x)
}
\arguments{
\item{x}{Any \R object.}
}
\value{
A non-negative integer.
}
\description{
Gets the length of an object without dispatching
}
\details{
This function returns \code{length(unclass(x))}, but tries to avoid
calling \code{unclass(x)} unless necessary.
}
\seealso{
\code{\link{.subset}()} and \code{\link{.subset2}()}.
}
\keyword{internal}
