% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-immediateCondition.R
\name{readImmediateConditions}
\alias{readImmediateConditions}
\title{Read All 'immediateCondition' RDS Files}
\usage{
readImmediateConditions(
  path = immediateConditionsPath(),
  pattern = "[.]rds$",
  include = getOption("future.relay.immediate", "immediateCondition"),
  remove = TRUE
)
}
\arguments{
\item{path}{(character string) The folder where the RDS files are.}

\item{pattern}{(character string) A regular expression selecting
the RDS files to be read.}

\item{include}{(character vector) The class or classes of the objects
to be kept.}
}
\value{
Returns a \link[base:list]{base::list} of immediateCondition objects.
}
\description{
Read All 'immediateCondition' RDS Files
}
\keyword{internal}
