% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api-reset.R
\name{reset}
\alias{reset}
\title{Reset a finished, failed, or interrupted future to a lazy future}
\usage{
reset(x, ...)
}
\arguments{
\item{x}{A Future.}

\item{\ldots}{Not used.}
}
\value{
\code{reset()} returns a lazy, vanilla \link{Future} that can be relaunched.
Resetting a running future results in a \link{FutureError}.
}
\description{
A future that has successfully completed, has been interrupted, or
failed due to an error, can be relaunched after resetting it.
}
\examples{
## Like mean(), but fails 90\% of the time
shaky_mean <- function(x) {
  if (as.double(Sys.time()) \%\% 1 < 0.90) stop("boom")
  mean(x)
}

x <- rnorm(100)

## Calculate the mean of 'x' with a risk of failing randomly
f <- future({ shaky_mean(x) })

## Relaunch until success
repeat({
  v <- tryCatch(value(f), error = identity)
  if (!inherits(v, "error")) break
  message("Resetting failed future, and retry in 0.1 seconds")
  f <- reset(f)
  Sys.sleep(0.1)
})
cat("mean:", v, "\n")
}
