% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.plan.R
\name{resetWorkers}
\alias{resetWorkers}
\title{Free up active background workers}
\usage{
resetWorkers(x, ...)
}
\arguments{
\item{x}{A FutureStrategy.}

\item{\ldots}{Not used.}
}
\description{
Free up active background workers
}
\details{
This function will resolve any active futures that is currently
being evaluated on background workers.
}
\examples{
resetWorkers(plan())

}
\keyword{internal}
