% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resignalConditions.R
\name{resignalConditions}
\alias{resignalConditions}
\title{Signals Captured Conditions}
\usage{
resignalConditions(future, include = "condition", exclude = NULL, ...)
}
\arguments{
\item{future}{A resolved \link{Future}.}

\item{include}{A character string of \link[base:condition]{condition}
classes to signal.}

\item{exclude}{A character string of \link[base:condition]{condition}
classes \emph{not} to signal.}

\item{\ldots}{Not used.}
}
\value{
Returns the \code{future} object invisibly.
}
\description{
Captured conditions that meet the \code{include} and \code{exclude}
requirements are signaled \emph{in the order as they were captured}.
}
\seealso{
Conditions are signaled by
\code{\link[base:signalCondition]{signalCondition}()}.
}
