% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Future-class.R
\name{resolved.Future}
\alias{resolved}
\alias{resolved.Future}
\title{Check whether a future is resolved or not}
\usage{
\method{resolved}{Future}(future, ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{...}{Not used}
}
\value{
TRUE if future is resolved and FALSE if unresolved.
}
\description{
Check whether a future is resolved or not
}
\details{
This method needs to be implemented by the class that implement
the Future API.  The implementation must never throw an error,
but only return either TRUE or FALSE.
It should also be possible to use the method for polling the
future until it is resolved (without having to wait infinitly long),
e.g. \code{while (!resolved(future)) Sys.sleep(5)}.
}

