% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Future-class.R
\name{result.Future}
\alias{result.Future}
\alias{result}
\title{Get the results of a resolved future}
\usage{
\method{result}{Future}(future, ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{\dots}{Not used.}
}
\value{
The \link{FutureResult} object.
}
\description{
Get the results of a resolved future
}
\details{
This function is only part of the _backend_ Future API.
This function is _not_ part of the frontend Future API.
}
\keyword{internal}
