% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Future-class.R
\name{run.Future}
\alias{run.Future}
\alias{run}
\title{Run a future}
\usage{
\method{run}{Future}(future, ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{\dots}{Not used.}
}
\value{
The \link{Future} object.
}
\description{
Run a future
}
\details{
This function can only be called once per future.
Further calls will result in an informative error.
If a future is not run when its value is queried,
then it is run at that point.
}
