% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-immediateCondition.R
\name{save_rds}
\alias{save_rds}
\title{Robustly Saves an Object to RDS File Atomically}
\usage{
save_rds(object, pathname, ...)
}
\arguments{
\item{object}{The \R object to be save.}

\item{pathname}{RDS file to written.}

\item{\ldots}{(optional) Additional arguments passed to \code{\link[base:readRDS]{base::saveRDS()}}.}
}
\value{
(invisible) The pathname of the RDS written.
}
\description{
Robustly Saves an Object to RDS File Atomically
}
\details{
Uses \link[base:readRDS]{base::saveRDS} internally but writes the object atomically by first
writing to a temporary file which is then renamed.
}
\keyword{internal}
