% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplifies a list like object}
\usage{
simplify(x, ...)
}
\arguments{
\item{x}{A list or a list environment}

\item{...}{Passed to \code{unlist(x)}.}
}
\value{
Returns a simplified object.
}
\description{
Simplifies a list like object
}

