% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky_globals.R
\name{sticky_globals}
\alias{sticky_globals}
\title{Place a sticky-globals environment immediately after the global environment}
\usage{
sticky_globals(erase = FALSE, name = "future:sticky_globals", pos = 2L)
}
\arguments{
\item{erase}{(logical) If TRUE, the environment is erased, otherwise not.}

\item{name}{(character) The name of the environment on the \link[base:search]{base::search}
path.}

\item{pos}{(integer) The position on the search path where the
environment should be positioned.  If \code{pos == 0L}, then the environment
is detached, if it exists.}
}
\value{
(invisible; environment) The environment.
}
\description{
Place a sticky-globals environment immediately after the global environment
}
\keyword{internal}
