% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{summary.FutureJournal}
\alias{summary.FutureJournal}
\title{Summarize Timing and Memory Profiling}
\usage{
\method{summary}{FutureJournal}(object, by = c("evaluation", "process", "future"), ...)
}
\arguments{
\item{by}{If \code{"evaluation"} (default), summarize across all evaluations.
Each evaluation is resolved by one future.
If \code{"process"}, summarize journal statistics per process,
which includes the main R session, and all workers (if any).}
}
\value{
A data.frame of class \code{FutureJournalSummary}.
}
\description{
\emph{WARNING: This function is under development. It can change at any time.
For now, please, do not depend on this function in a published R package.}
}
