% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{supportsResources}
\alias{supportsResources}
\title{Check If Future Strategy Supports Specified Resources}
\usage{
supportsResources(strategy, resources, ...)
}
\arguments{
\item{strategy}{A future strategy class \code{future}.}

\item{resources}{A resources formula or NULL.}

\item{\ldots}{For internal use only.}
}
\value{
(logical) TRUE if \code{resources} are supported, otherwise FALSE.
}
\description{
Check If Future Strategy Supports Specified Resources
}
\details{
This function is intended for internal use and developing additional
future backends.
}
