% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate.R
\name{terminate}
\alias{terminate}
\alias{terminate.Future}
\alias{terminate.MulticoreFuture}
\alias{terminate.MultisessionFuture}
\alias{terminate.SOCKnode}
\alias{terminate.SOCKcluster}
\alias{terminate.integer}
\title{Terminate a Future Process}
\usage{
terminate(x, ...)

\method{terminate}{Future}(x, ...)

\method{terminate}{MulticoreFuture}(x, ...)

\method{terminate}{MultisessionFuture}(x, ...)

\method{terminate}{SOCKnode}(x, ...)

\method{terminate}{SOCKcluster}(x, ...)

\method{terminate}{integer}(x, signal = tools::SIGINT, ...)
}
\arguments{
\item{x}{A future to be terminated.}

\item{signal}{Integer specifying the signal level used to terminate
the future.}

\item{\ldots}{(optional) Reserved for future use.}
}
\value{
Returns (invisibly) a logical indicating success or failure of
process signalling.
}
\description{
Terminate a Future Process
}
