% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak}
\alias{tweak}
\title{Tweaks a future function by adjusting its default arguments}
\usage{
tweak(strategy, ..., penvir = parent.frame())
}
\arguments{
\item{strategy}{An existing future function or the name of one.}

\item{\dots}{Named arguments to replace the defaults of existing
arguments.}

\item{penvir}{The environment used when searching for a future
function by its name.}
}
\value{
a future function.
}
\description{
Tweaks a future function by adjusting its default arguments
}
\seealso{
Use \code{\link{plan}()} to set a future to become the
new default strategy.
}
