% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Future-class.R
\name{value.Future}
\alias{value}
\alias{value.Future}
\title{Gets the value of a future}
\usage{
\method{value}{Future}(future, onCondition = c("signal", "return"), ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{onCondition}{A character string specifying how
\link[base]{conditions} should be handled in case they occur.
If \code{"signal"}, the condition is generated, e.g. captured
errors and warnings are \emph{"signaled"} (re-thrown).
If \code{"return"}, they are instead \emph{returned} as is.}

\item{...}{Not used.}
}
\value{
An R object of any data type.
}
\description{
Gets the value of a future.  If the future is unresolved, then
the evaluation blocks until the future is resolved.
}
\details{
This method needs to be implemented by the class that implement
the Future API.
}

