% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api-with.R
\name{with.FutureStrategyList}
\alias{with.FutureStrategyList}
\title{Evaluate an expression using a temporarily set future plan}
\usage{
\method{with}{FutureStrategyList}(data, expr, ...)
}
\arguments{
\item{data}{The future plan to use temporarily.}

\item{The}{R expression to be evaluated.}

\item{\ldots}{Not used.}
}
\value{
The value of the expression evaluated.
}
\description{
Evaluate an expression using a temporarily set future plan
}
\examples{
# Evaluate a future using the 'multisession' plan
with(plan(multisession, workers = 2), {
  f <- future(Sys.getpid())
  w_pid <- value(f)
  c(main = Sys.getpid(), worker = w_pid)
})

}
