% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrace.R
\name{backtrace}
\alias{backtrace}
\title{Back trace the expressions evaluated before a condition was caught}
\usage{
backtrace(future, envir = parent.frame(), ...)
}
\arguments{
\item{future}{The future with a caught condition.}

\item{envir}{the environment where to locate the future.}

\item{\dots}{Not used.}
}
\value{
A list of calls.
}
\description{
Back trace the expressions evaluated before a condition was caught
}
