% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI.R
\name{BMI}
\alias{BMI}
\title{BMI}
\usage{
BMI(weight, growth, sex = c("female", "male"))
}
\arguments{
\item{weight}{numeric, weight of the person for whom we calculate the BMI
expressed in kg}

\item{growth}{numeric, growth of the person for whom we calculate the BMI
expressed in meters}

\item{sex}{string, sex of the person for whom we calculate BMI. Possible
values for the parameter: \code{female}, \code{male}}
}
\description{
Function for calculating BMI.
}
\examples{
BMI(100, 1.7)
BMI(85, 1.9, "female")

}
