% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api-cancel.R
\name{cancel}
\alias{cancel}
\title{Cancel a future}
\usage{
cancel(x, interrupt = TRUE, ...)
}
\arguments{
\item{x}{A Future.}

\item{interrupt}{If TRUE, running futures are interrupted, if possible.}

\item{\ldots}{All arguments used by the S3 methods.}
}
\value{
\code{cancel()} returns the \link{Future} flagged as "canceled".

Canceling a lazy or a finished future has no effect.
}
\description{
Cancels futures, with the option to interrupt running ones.
}
