% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFuture-class.R
\name{ClusterFuture-class}
\alias{ClusterFuture}
\alias{ClusterFuture-class}
\alias{MultisessionFuture}
\alias{MultisessionFuture-class}
\title{A cluster future is a future whose value will be resolved asynchroneously in a parallel process}
\usage{
ClusterFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  local = !persistent, gc = !persistent, persistent = FALSE,
  workers = NULL, ...)
}
\arguments{
\item{expr}{An R \link[base]{expression}.}

\item{envir}{The \link{environment} in which the evaluation
is done (or inherits from if \code{local} is TRUE).}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{local}{If TRUE, the expression is evaluated such that
all assignments are done to local temporary environment, otherwise
the assignments are done in the global environment of the cluster node.}

\item{gc}{If TRUE, the garbage collector run after the future
is resolved (in the process that evaluated the future).}

\item{persistent}{If FALSE, the evaluation environment is cleared
from objects prior to the evaluation of the future.}

\item{workers}{A \code{\link[parallel:makeCluster]{cluster}}.
Alternatively, a character vector of host names or a numeric scalar,
for creating a cluster via \code{\link[parallel]{makeCluster}(workers)}.}

\item{\dots}{Additional named elements of the future.}
}
\value{
An object of class \code{ClusterFuture}.
}
\description{
A cluster future is a future whose value will be resolved asynchroneously in a parallel process
}
\seealso{
To evaluate an expression using "cluster future", see function
\code{\link{cluster}()}.
}
\keyword{internal}

