% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantFuture-class.R
\name{ConstantFuture-class}
\alias{ConstantFuture-class}
\alias{ConstantFuture}
\title{A future with a constant value}
\usage{
ConstantFuture(..., globals = TRUE, envir = emptyenv())
}
\arguments{
\item{...}{Additional named elements of the future.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[=future]{future()}}.}

\item{envir}{The \link{environment} from where global objects should be
identified.}
}
\value{
\code{ConstantFuture()} returns an object of class \code{ConstantFuture}.
}
\description{
A constant future is a future whose expression is a constant
and therefore by definition is already resolved upon creation.
}
\keyword{internal}
