% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantFuture-class.R
\name{ConstantFuture-class}
\alias{ConstantFuture}
\alias{ConstantFuture-class}
\title{A future with a constant value}
\usage{
ConstantFuture(value = NULL, ...)
}
\arguments{
\item{value}{An R value.}

\item{\dots}{Not used.}
}
\value{
An object of class \code{ConstantFuture}.
}
\description{
A constant future is a future whose expression is a constant
and therefore already resolved upon creation
}
\keyword{internal}

