% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantFuture-class.R
\name{ConstantFuture-class}
\alias{ConstantFuture-class}
\alias{ConstantFuture}
\title{A future with a constant value}
\usage{
ConstantFuture(expr = NULL, envir = emptyenv(), substitute = FALSE,
  local = FALSE, ...)
}
\arguments{
\item{expr}{An R \link[base]{expression}.}

\item{envir}{The \link{environment} in which the evaluation
is done (or inherits from if \code{local} is TRUE).}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{local}{If TRUE, the expression is evaluated such that
all assignments are done to local temporary environment, otherwise
the assignments are done in the calling environment.}

\item{\dots}{Not used.}
}
\value{
An object of class \code{ConstantFuture}.
}
\description{
A constant future is a future whose expression is a constant
and therefore by definition already resolved upon creation.
}
\keyword{internal}
