% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EagerFuture-class.R
\name{EagerFuture-class}
\alias{EagerFuture}
\alias{EagerFuture-class}
\title{An eager future is a future whose value will be resolved immediately}
\usage{
EagerFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  globals = TRUE, local = TRUE, ...)
}
\arguments{
\item{...}{Additional named elements passed to \code{\link{Future}()}.}
}
\value{
An object of class \code{EagerFuture}.
}
\description{
An eager future is a future whose value will be resolved immediately
}
\seealso{
To evaluate an expression using "eager future", see function
\code{\link{eager}()}.
}
\keyword{internal}

