% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Future-class.R
\name{Future-class}
\alias{Future}
\alias{Future-class}
\title{A future represents a value that will be available at some point in the future}
\usage{
Future(object, ...)
}
\arguments{
\item{object}{An R object of a class that implements the Future API.}

\item{...}{Not used.}
}
\value{
An object of class \code{Future}.
}
\description{
A \emph{future} is an abstraction for a \emph{value} that may
available at some point in the future.  A future can either be
\code{unresolved} or \code{resolved}, a state which can be checked
with \code{\link{resolved}()}.  As long as it is \emph{unresolved}, the
value is not available.  As soon as it is \emph{resolved}, the value
is available via \code{\link{value}()}.
}
\seealso{
One function that creates a Future is \code{\link{future}()}.
It returns a Future that evaluates an R expression in the future.
An alternative approach is to use the \code{\link{\%<=\%}} infix
assignment operator, which creates a future from the
right-hand-side (RHS) R expression and assigns its future value
to a variable as a \emph{\link[base]{promise}}.
}

