% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Future-class.R
\name{Future-class}
\alias{Future}
\alias{Future-class}
\title{A future represents a value which may become available in the future}
\usage{
Future(object, ...)
}
\arguments{
\item{object}{An R object of a class that implements the Future API.}

\item{...}{Not used.}
}
\value{
An object of class \code{Future}.
}
\description{
A \emph{future} is an abstraction for a \emph{value} that may
available at some point in the future.  A future can either be
\code{unresolved} or \code{resolved}, which can be checked with
\code{isResolved()}.  As long as it is \emph{unresolved}, the value
is not available.  As soon as it is \emph{resolved}, the value is
available, which can take the form of an object of any data type
or a \link[base]{condition}.
}
\seealso{
One function that creates a Future is \code{\link{future}()}.
It returns a Future that evaluates an R expression in the future.
An alternative approach is to use the \code{\link{\%<=\%}} infix
operator, which asynchroneously evaluates an R expression and returns
a \emph{\link[base]{promise}} to assign the value to a variable, if
requested.
}

