% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_lapply.R
\name{future_lapply}
\alias{future_lapply}
\title{[DEFUNCT] Apply a Function over a List or Vector via Futures}
\usage{
future_lapply(...)
}
\arguments{
\item{\dots}{Not used.}
}
\description{
\emph{
IMPORTANT: The \code{future_lapply()} function has been moved to the
\href{https://cran.r-project.org/package=future.apply}{future.apply} package.
The version in this package (future) is defunct and gives an error if used.
It will be fully removed in an upcoming release.
Please update your code to make use of
\code{future.apply::future_lapply()} instead.
}
}
\keyword{internal}
