% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/futureAssign.R
\name{futureAssign}
\alias{\%<=\%}
\alias{\%=>\%}
\alias{futureAssign}
\title{Future evaluation and assignment}
\usage{
futureAssign(name, value, envir = parent.frame(), assign.env = envir,
  substitute = TRUE)
}
\arguments{
\item{name}{the name of the variable (and the future) to assign.}

\item{value}{the R expression to be evaluated in the future and
whose value will be assigned to the variable.}

\item{envir}{The environment from which global variables used by
the expression should be search for.}

\item{assign.env}{The environment to which the variable should
be assigned.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}
}
\value{
A \code{\link[future]{Future}} invisibly.
}
\description{
Method and infix operators for future assignments of the value
of an expression which is evaluated in the future.
}
\details{
This function creates a future and a corresponding
"\emph{\link[base]{promise}}", which hold the future's value.
Both the future and the promise are assigned to environment
\code{assign.env}.  The name of the promise is given by \code{name}
and the name of the future is \code{.future_<name>}.
The future is also returned invisibly.
}
\examples{
## A global variable
a <- 1

## Three future evaluations
A \%<=\% { x <- 0.1 }
B \%<=\% { y <- 0.2 }
C \%<=\% { z <- a+0.3 }

## Sleep until 'C' is available
message("Value 'C': ", C)

## Sleep until 'A' is available
message("Value 'A': ", A)

## Sleep until 'C' is available
message("Value 'B': ", B)
}

