% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureCondition-class.R
\name{FutureCondition}
\alias{FutureCondition}
\alias{FutureMessage}
\alias{FutureWarning}
\alias{FutureError}
\alias{RngFutureCondition}
\alias{RngFutureWarning}
\alias{RngFutureError}
\alias{UnexpectedFutureResultError}
\title{A condition (message, warning, or error) that occurred while orchestrating a future}
\usage{
FutureCondition(message, call = NULL, future = NULL)

FutureMessage(message, call = NULL, future = NULL)

FutureWarning(message, call = NULL, future = NULL)

FutureError(message, call = NULL, future = NULL)

RngFutureCondition(message = NULL, call = NULL, future)

RngFutureWarning(...)

RngFutureError(...)

UnexpectedFutureResultError(future, hint = NULL)
}
\arguments{
\item{message}{A message condition.}

\item{call}{The call stack that led up to the condition.}

\item{future}{The \link{Future} involved.}

\item{hint}{(optional) A string with a suggestion on what might be wrong.}
}
\value{
An object of class FutureCondition which inherits from class
\link[base:conditions]{condition} and FutureMessage, FutureWarning,
and FutureError all inherits from FutureCondition.
Moreover, a FutureError inherits from \link[base:conditions]{error},
a FutureWarning from \link[base:conditions]{warning}, and
a FutureMessage from \link[base:conditions]{message}.
}
\description{
While \emph{orchestrating} (creating, launching, querying, collection)
futures, unexpected run-time errors (and other types of conditions) may
occur.  Such conditions are coerced to a corresponding FutureCondition
class to help distinguish them from conditions that occur due to the
\emph{evaluation} of the future.
}
\keyword{internal}
