% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureResult-class.R
\name{FutureResult}
\alias{FutureResult}
\alias{as.character.FutureResult}
\title{Results from resolving a future}
\usage{
FutureResult(value = NULL, condition = NULL, calls = NULL, ...,
  version = "1.7")

\method{as.character}{FutureResult}(x, ...)
}
\arguments{
\item{value}{The value of the future expression.
If the expression was not fully resolved (e.g. an error) occurred,
the the value is `NULL`.}

\item{condition}{A [[base::condition]] captured while resolving the future,
if any.  This is typically an error.}

\item{calls}{A list of calls that led up to the captured condition, if any.}

\item{version}{The version format of the results.}

\item{\ldots}{(optional) Additional named results to be returned.}
}
\value{
An object of class FutureResult.
}
\description{
Results from resolving a future
}
\details{
This function is only part of the _backend_ Future API.
This function is _not_ part of the frontend Future API.
}
\keyword{internal}
