% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{futures}
\alias{futures}
\title{Gets all futures in an object}
\usage{
futures(x, ...)
}
\arguments{
\item{x}{An environment, a list, or a list environment.}

\item{\dots}{Not used.}
}
\value{
An object of same type as \code{x} and with the same names
and/or dimensions, if set.
}
\description{
Gets all futures in an environment, a list, or a list environment
and returns an object of the same class (and dimensions).
Non-future elements are returned as is.
}
\details{
This function is useful for retrieve futures that were created via
future assignments (\code{\%<-\%}) and therefore stored as promises.
This function turns such promises into standard \code{Future}
objects.
}
