% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureSessionInfo.R
\name{futureSessionInfo}
\alias{futureSessionInfo}
\title{Get future-specific session information and validate current backend}
\usage{
futureSessionInfo(test = TRUE)
}
\arguments{
\item{test}{If TRUE, one or more futures are created to query workers
and validate their information.}
}
\value{
Nothing.
}
\description{
Get future-specific session information and validate current backend
}
\examples{
plan(multisession, workers = 2)
futureSessionInfo()
plan(sequential)
}
