% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_OP.R
\name{\%globals\%}
\alias{\%globals\%}
\alias{\%packages\%}
\title{Specify globals and packages for a future assignment}
\usage{
fassignment \%globals\% globals
fassignment \%packages\% packages
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% { expr }}.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[=future]{future()}}.}

\item{packages}{(optional) a character vector specifying packages
to be attached in the \R environment evaluating the future.}
}
\description{
Specify globals and packages for a future assignment
}
