% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_OP.R
\name{\%globals\%}
\alias{\%globals\%}
\title{Specify globals for a future assignment}
\usage{
fassignment \%globals\% globals
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{globals}{(optional) a logical, a character vector,
or a named list for controlling how globals are handled.
For details, see section 'Globals used by future expressions'
in the help for \code{\link{future}()}.}
}
\description{
Specify globals for a future assignment
}
\details{
\code{a \%globals\% b} is short for \code{a \%tweak\% list(globals = b)}.
}
