% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_OP.R
\name{\%lazy\%}
\alias{\%lazy\%}
\title{Control lazy / eager evaluation for a future assignment}
\usage{
fassignment \%lazy\% lazy
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{lazy}{If \code{FALSE} (default), the future is resolved eagerly
(immediately), otherwise not.}
}
\description{
Control lazy / eager evaluation for a future assignment
}
