% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_OP.R
\name{\%seed\%}
\alias{\%seed\%}
\title{Set random seed for future assignment}
\usage{
fassignment \%seed\% seed
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{seed}{(optional) A L'Ecuyer-CMRG RNG seed.}
}
\description{
Set random seed for future assignment
}
