% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal}
\alias{journal}
\title{Gets a journal of events for a future}
\usage{
journal(x, ...)
}
\arguments{
\item{x}{A \link{Future} object.}

\item{\ldots}{Not used.}
}
\value{
A data frame with columns \code{step}, \code{start}, \code{stop}, and \code{duration}.a
}
\description{
Gets a journal of events for a future
}
