% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal}
\alias{journal}
\title{Gets a journal of events for a future}
\usage{
journal(x, ...)
}
\arguments{
\item{x}{A \link{Future} object.}

\item{\ldots}{Not used.}
}
\value{
A data frame of class \code{FutureJournal} with columns:
\enumerate{
\item \code{event} (character string)
\item \code{start} (POSIXct)
\item \code{at} (difftime)
\item \code{duration} (difftime)
\item \code{future_label} (character string)
\item \code{future_uuid} (character string)
\item \code{session_uuid} (character string)
}

Possible events are:
\itemize{
\item \code{create} - the future was created
\item \code{launch} - the future was launched (aka run)
\item \code{evaluate} - the future was evaluated
\item \code{resolved} - the future was queried (may be occur multiple times)
\item \code{gather} - the results was retrieved
}

The data frame is sorted by the \code{at} time.
Note that the timestamps for the \code{evaluate} event are based on the local
time on the worker. The system clocks on the worker and the calling R
system may be out of sync.
}
\description{
Gets a journal of events for a future
}
\examples{
library(future)
journal <- future:::journal

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)

## Stop parallel workers and undo journaling settings
plan(sequential)
options(oopts)
}
