% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mainModule}
\alias{mainModule}
\title{Server part of module for calculating energy of meal}
\usage{
mainModule(input, output, session, logger_threshold = "INFO")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{logger_threshold}{threshold for logger}
}
\description{
Module for adding products and their weights to the meal.
}
