% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelbrot.R
\name{mandelbrot}
\alias{as.raster.Mandelbrot}
\alias{mandelbrot}
\alias{plot.Mandelbrot}
\title{Mandelbrot convergence counts}
\usage{
mandelbrot(x = -0.75, y = 0, side = 2.5, resolution = 400L,
  maxIter = 200L, tau = 2)
}
\arguments{
\item{x, y, side}{Center location and the length of sides of the
complex plane for which convergence counts
should be calculated.}

\item{resolution}{Number of bins along each axis.}

\item{maxIter}{Maximum number of iterations per bin.}

\item{tau}{A threshold; the radius when calling
divergence (Mod(z) > tau).}
}
\value{
Returns an integer matrix (of class Mandelbrot) with
}
\description{
Mandelbrot convergence counts
}
\examples{
counts <- mandelbrot()
plot(counts)
}
\author{
This \code{mandelbrot()} function was inspired by and
adopted from similar GPL code of Martin Maechler (available
from ftp://stat.ethz.ch/U/maechler/R/ on 2005-02-18 [sic!]).
}
\keyword{internal}

