% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulticoreFuture-class.R
\name{MulticoreFuture-class}
\alias{MulticoreFuture-class}
\alias{MulticoreFuture}
\title{An multicore future is a future whose value will be resolved asynchronously in a parallel process}
\usage{
MulticoreFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  globals = TRUE,
  lazy = FALSE,
  ...
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[=future]{future()}}.}

\item{lazy}{If FALSE (default), the future is resolved
eagerly (starting immediately), otherwise not.}

\item{...}{Additional named elements passed to \code{\link[=Future]{Future()}}.}
}
\value{
An object of class \code{MulticoreFuture}.
}
\description{
An multicore future is a future whose value will be resolved asynchronously in a parallel process
}
\seealso{
To evaluate an expression using "multicore future", see function
\code{\link[=multicore]{multicore()}}.
}
\keyword{internal}
