% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulticoreFuture.R
\name{MulticoreFuture-class}
\alias{MulticoreFuture}
\alias{MulticoreFuture-class}
\title{An multicore future is a future whose value will be resolved asynchroneously in a parallel process}
\usage{
MulticoreFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  ...)
}
\arguments{
\item{object}{An R \link[base]{environment}.}

\item{\dots}{Not used.}
}
\value{
An object of class \code{MulticoreFuture}.
}
\description{
An multicore future is a future whose value will be resolved asynchroneously in a parallel process
}
\seealso{
To evaluate an expression using "multicore future", see function
\code{\link{multicore}()}.
}
\keyword{internal}

