% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulticoreFuture-class.R
\name{MulticoreFuture-class}
\alias{MulticoreFuture-class}
\alias{MulticoreFuture}
\title{A multicore future is a future whose value will be resolved asynchronously in a parallel process}
\usage{
MulticoreFuture(expr = NULL, substitute = TRUE, envir = parent.frame(), ...)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{...}{Additional named elements passed to \code{\link[=Future]{Future()}}.}
}
\value{
\code{MulticoreFuture()} returns an object of class \code{MulticoreFuture}.
}
\description{
A multicore future is a future whose value will be resolved asynchronously in a parallel process
}
\section{Usage}{

To use 'multicore' futures, use \code{plan(multicore, ...)}, cf. \link{multicore}.
}

\keyword{internal}
