% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulticoreFuture-class.R
\name{MulticoreFuture-class}
\alias{MulticoreFuture}
\alias{MulticoreFuture-class}
\title{An multicore future is a future whose value will be resolved asynchroneously in a parallel process}
\usage{
MulticoreFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  globals = TRUE, ...)
}
\arguments{
\item{...}{Additional named elements passed to \code{\link{Future}()}.}
}
\value{
An object of class \code{MulticoreFuture}.
}
\description{
An multicore future is a future whose value will be resolved asynchroneously in a parallel process
}
\seealso{
To evaluate an expression using "multicore future", see function
\code{\link{multicore}()}.
}
\keyword{internal}

