% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocess.R
\name{multiprocess}
\alias{multiprocess}
\title{Create a multiprocess future whose value will be resolved asynchronously using multicore or a multisession evaluation}
\usage{
multiprocess(..., workers = availableCores(), envir = parent.frame())
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link[=Future]{Future()}}.}

\item{workers}{Ignored in \strong{future} (>= 1.31.0).}

\item{envir}{The \link{environment} from where global objects should be
identified.}
}
\value{
A \link{SequentialFuture} (sic!) since \strong{future} 1.31.0.
}
\description{
\strong{WARNING: The 'multiprocess' future plan is deprecated.
Instead, explicitly specify 'multisession' or 'multicore'.  The former works
everywhere and is the recommended one between the two. \emph{Forked processing},
which 'multicore' uses, is unstable in various environment and setups.
The 'multiprocess' alias is therefore being phased out, and is now
equal to using 'sequential' (sic!)}
}
\keyword{internal}
