% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrOfWorkers.R
\name{nbrOfWorkers}
\alias{nbrOfWorkers}
\title{Get the number of workers available}
\usage{
nbrOfWorkers(evaluator = NULL)
}
\arguments{
\item{evaluator}{A future evaluator function.
If NULL (default), the current evaluator as returned
by \code{\link[=plan]{plan()}} is used.}
}
\value{
A positive number in \eqn{{1, 2, 3, ...}}.
Note, it may also be \code{+Inf} for certain types of backends.
}
\description{
Get the number of workers available
}
\examples{
plan(multisession)
nbrOfWorkers()  ## == availableCores()

plan(sequential)
nbrOfWorkers()  ## == 1
}
