% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullcon}
\alias{nullcon}
\title{Creates a connection to the system null device}
\usage{
nullcon()
}
\value{
Returns a open, binary \code{\link[base:connections]{base::connection()}}.
}
\description{
Creates a connection to the system null device
}
\keyword{internal}
