% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.re-exports.R
\name{re-exports}
\alias{re-exports}
\alias{as.cluster}
\alias{autoStopCluster}
\alias{availableCores}
\alias{availableWorkers}
\alias{makeClusterMPI}
\alias{makeClusterPSOCK}
\alias{makeNodePSOCK}
\alias{supportsMulticore}
\title{Functions Moved to 'parallelly'}
\description{
The following function used to be part of \pkg{future} but has since
been migrated to \pkg{parallelly}.  The migration started with
\pkg{future} 1.20.0 (November 2020).  They were moved because they
are also useful outside of the \pkg{future} framework.
}
\details{
\itemize{
\item \code{\link[parallelly:as.cluster]{parallelly::as.cluster()}}
\item \code{\link[parallelly:autoStopCluster]{parallelly::autoStopCluster()}}
\item \code{\link[parallelly:availableCores]{parallelly::availableCores()}}
\item \code{\link[parallelly:availableWorkers]{parallelly::availableWorkers()}}
\item \code{\link[parallelly:makeClusterMPI]{parallelly::makeClusterMPI()}}
\item \code{\link[parallelly:makeClusterPSOCK]{parallelly::makeClusterPSOCK()}}
\item \code{\link[parallelly:makeClusterPSOCK]{parallelly::makeNodePSOCK()}}
\item \code{\link[parallelly:supportsMulticore]{parallelly::supportsMulticore()}}
}

For backward-compatible reasons, these functions remain available as
exact copies also from this package (as re-exports).  For example,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cl <- parallelly::makeClusterPSOCK(2)
}\if{html}{\out{</div>}}

can still be accessed as:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cl <- future::makeClusterPSOCK(2)
}\if{html}{\out{</div>}}
}
