% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protected_api-result.R
\name{result.Future}
\alias{result.Future}
\alias{result}
\title{Get the results of a resolved future}
\usage{
\method{result}{Future}(future, ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{\ldots}{Not used.}
}
\value{
The \link{FutureResult} object.
It may signal a \link{FutureError}, if there is a significant orchestration
error. For example, if the parallel worker process terminated abruptly
("crashed"), then a \link{FutureInterruptError} is signaled.
}
\description{
Get the results of a resolved future
}
\details{
This function is only part of the \emph{backend} Future API.
This function is \emph{not} part of the frontend Future API.
}
\keyword{internal}
