% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionDetails.R
\name{sessionDetails}
\alias{sessionDetails}
\title{Outputs details on the current R session}
\usage{
sessionDetails(env = FALSE)
}
\arguments{
\item{env}{If TRUE, \code{Sys.getenv()} information is returned.}
}
\value{
Invisibly a list of all details.
}
\description{
Outputs details on the current R session
}
\details{
Note that remote futures use \code{persistent=TRUE} by default.
}
\keyword{internal}
