% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multicore.R
\name{supportsMulticore}
\alias{supportsMulticore}
\title{Check whether multicore processing is supported or not}
\usage{
supportsMulticore()
}
\value{
TRUE if multicore processing is supported, otherwise FALSE.
}
\description{
Multicore futures are only supported on systems supporting
multicore processing.  R supports this on most systems,
except on the Microsoft Windows.
}
\seealso{
To use multicore futures, use \code{\link{plan}(\link{multicore})}.
}

