% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak_parallel_PSOCK.R
\name{tweak_parallel_PSOCK}
\alias{tweak_parallel_PSOCK}
\title{Tweak PSOCK backend of the parallel package}
\usage{
tweak_parallel_PSOCK(user = TRUE, revtunnel = TRUE, rshopts = TRUE,
  reset = FALSE)
}
\arguments{
\item{user}{If TRUE, parallel is tweaked to only pass username to ssh if it is specified via argument \code{user}.}

\item{revtunnel}{If TRUE, parallel is tweaked to make use of reverse SSH tunneling.}

\item{rshopts}{A character vector of additional command-line options passed to the \code{rshcmd} executable.}

\item{reset}{If TRUE, all tweaks are undone.}
}
\value{
Nothing.
}
\description{
Tweak PSOCK backend of the parallel package
}
\references{
\url{https://github.com/HenrikBengtsson/Wishlist-for-R/issues/32}\cr
}
\keyword{internal}

