% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_api-UniprocessFuture-class.R
\name{UniprocessFuture-class}
\alias{UniprocessFuture-class}
\alias{UniprocessFuture}
\title{An uniprocess future is a future whose value will be resolved synchronously in the current process}
\usage{
UniprocessFuture(expr = NULL, substitute = TRUE, envir = parent.frame(), ...)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{\ldots}{Additional named elements passed to \code{\link[=Future]{Future()}}.}
}
\value{
\code{UniprocessFuture()} returns an object of class \code{UniprocessFuture}.
}
\description{
An uniprocess future is a future whose value will be resolved synchronously in the current process
}
\keyword{internal}
