% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Future-class.R
\name{value.Future}
\alias{value.Future}
\alias{value}
\title{The value of a future}
\usage{
\method{value}{Future}(future, signal = TRUE, ...)
}
\arguments{
\item{future}{A \link{Future}.}

\item{signal}{A logical specifying whether (\link[base]{conditions})
should signaled or be returned as values.}

\item{\dots}{Not used.}
}
\value{
An R object of any data type.
}
\description{
Gets the value of a future.  If the future is unresolved, then
the evaluation blocks until the future is resolved.
}
\details{
This method needs to be implemented by the class that implement
the Future API.
}
