% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Gets all values in an object}
\usage{
values(x, ...)
}
\arguments{
\item{x}{An environment, a list, or a list environment.}

\item{\dots}{Additional arguments passed to \code{value()} of each future.}
}
\value{
An object of same type as \code{x} and with the same names
and/or dimensions, if set.
}
\description{
Gets all values in an environment, a list, or a list environment
and returns an object of the same class (and dimensions).
All future elements are replaced by their corresponding
\code{value()} values.  For all other elements, the existing
object is kept.
}
